+++ saleor/graphql/attribute/tests/mutations/test_attribute_value_create.py
@@ -27,9 +27,9 @@ def test_validate_value_is_unique(color_attribute):
 
 CREATE_ATTRIBUTE_VALUE_MUTATION = """
     mutation createAttributeValue(
-        $attributeId: ID!, $name: String!) {
+        $attributeId: ID!, $name: String!, $value: String) {
     attributeValueCreate(
-        attribute: $attributeId, input: {name: $name}) {
+        attribute: $attributeId, input: {name: $name, value: $value}) {
         attributeErrors {
             field
             message
@@ -38,6 +38,7 @@ CREATE_ATTRIBUTE_VALUE_MUTATION = """
         attribute {
             values {
                 name
+                value
             }
         }
         attributeValue {
@@ -57,7 +58,8 @@ def test_create_attribute_value(
     query = CREATE_ATTRIBUTE_VALUE_MUTATION
     attribute_id = graphene.Node.to_global_id("Attribute", attribute.id)
     name = "test name"
-    variables = {"name": name, "attributeId": attribute_id}
+    value = "test value"
+    variables = {"name": name, "value": value, "attributeId": attribute_id}
 
     # when
     response = staff_api_client.post_graphql(
@@ -73,6 +75,7 @@ def test_create_attribute_value(
     assert attr_data["name"] == name
     assert attr_data["slug"] == slugify(name)
     assert name in [value["name"] for value in data["attribute"]["values"]]
+    assert value in [value["value"] for value in data["attribute"]["values"]]
 
 
 def test_create_attribute_value_not_unique_name(